refactor: use fzf-lua instead of telescope

+++ nvim-fredrik/lua/fredrik/plugins/codecompanion.lua
@@ -1,17 +1,3 @@
--- add 2 commands:
---    CodeCompanionSave [space delimited args]
---    CodeCompanionLoad
--- Save will save current chat in a md file named 'space-delimited-args.md'
--- Load will use a telescope filepicker to open a previously saved chat and open it in a new chat buffer
-
--- create a folder to store our chats
-local Path = require("plenary.path")
-local data_path = vim.fn.stdpath("data") -- ~/.local/share/<NVIM_APPNAME>
-local save_folder = Path:new(data_path, "codecompanion_chats")
-if not save_folder:exists() then
-  save_folder:mkdir({ parents = true })
-end
-
 --- Anthropic config for CodeCompanion.
 local anthropic_fn = function()
   local anthropic_config = {
@@ -66,28 +52,31 @@ local supported_adapters = {
   ollama = ollama_fn,
 }
 
+-- add 2 commands:
+--    CodeCompanionSave [space delimited args]
+--    CodeCompanionLoad
+-- Save will save current chat in a md file named 'space-delimited-args.md'
+-- Load will use a telescope filepicker to open a previously saved chat and open it in a new chat buffer
+
+-- create a folder to store our chats
+local Path = require("plenary.path")
+local data_path = vim.fn.stdpath("data") -- ~/.local/share/<NVIM_APPNAME>
+local save_folder = Path:new(data_path, "codecompanion_chats")
+if not save_folder:exists() then
+  save_folder:mkdir({ parents = true })
+end
+
 -- telescope picker for our saved chats
 vim.api.nvim_create_user_command("CodeCompanionLoad", function()
-  local t_builtin = require("telescope.builtin")
-  local t_actions = require("telescope.actions")
-  local t_action_state = require("telescope.actions.state")
-  local t_pickers = require("telescope.pickers")
-  local t_finders = require("telescope.finders")
-  local t_conf = require("telescope.config").values
-
   local function select_adapter(filepath)
-    local picker = t_pickers.new({}, {
-      prompt_title = "Select CodeCompanion Adapter",
-      finder = t_finders.new_table({
-        results = vim.tbl_keys(supported_adapters),
-      }),
-      sorter = t_conf.generic_sorter({}),
-      attach_mappings = function(prompt_bufnr, _)
-        t_actions.select_default:replace(function()
-          t_actions.close(prompt_bufnr)
-          local selection = t_action_state.get_selected_entry()
-          local adapter = selection[1]
+    local fzf = require("fzf-lua")
+    local adapters = vim.tbl_keys(supported_adapters)
 
+    fzf.fzf_exec(adapters, {
+      prompt = "Select CodeCompanion Adapter> ",
+      actions = {
+        ["default"] = function(selected)
+          local adapter = selected[1]
           -- Open new CodeCompanion chat with selected adapter
           vim.cmd("CodeCompanionChat " .. adapter)
 
@@ -99,36 +88,37 @@ vim.api.nvim_create_user_command("CodeCompanionLoad", function()
 
           -- Paste contents into the new chat buffer
           vim.api.nvim_buf_set_lines(current_buf, 0, -1, false, lines)
-        end)
-        return true
-      end,
+        end,
+      },
     })
-    picker:find()
   end
 
   local function start_picker()
-    t_builtin.find_files({
-      prompt_title = "Saved CodeCompanion Chats | <c-d>: delete",
-      cwd = save_folder:absolute(),
-      attach_mappings = function(prompt_bufnr, map)
-        map("i", "<c-d>", function()
-          local selection = t_action_state.get_selected_entry()
-          local filepath = selection.path or selection.filename
-          os.remove(filepath)
-          t_actions.close(prompt_bufnr)
-          start_picker()
-        end)
+    local fzf = require("fzf-lua")
+    local files = vim.fn.glob(save_folder:absolute() .. "/*", false, true)
 
-        t_actions.select_default:replace(function()
-          local selection = t_action_state.get_selected_entry()
-          local filepath = selection.path or selection.filename
-          t_actions.close(prompt_bufnr)
-          select_adapter(filepath)
-        end)
-        return true
-      end,
+    fzf.fzf_exec(files, {
+      prompt = "Saved CodeCompanion Chats | <c-r>: remove >",
+      previewer = "builtin",
+      actions = {
+        ["default"] = function(selected)
+          if #selected > 0 then
+            local filepath = selected[1]
+            select_adapter(filepath)
+          end
+        end,
+        ["ctrl-r"] = function(selected)
+          if #selected > 0 then
+            local filepath = selected[1]
+            os.remove(filepath)
+            -- Refresh the picker
+            start_picker()
+          end
+        end,
+      },
     })
   end
+
   start_picker()
 end, {})
